requirements.txt
Django==3.1.1
line-bot-sdk==1.16.0
django-allauth==0.42.0
django-model-utils==4.0.0
djangorestframework==3.11.1
django-ipware==3.0.1
PyJWT==1.7.1
celery==4.4.7
Pillow==7.2.0
# gunicorn
# mysqlclient
